requirements.txt
discord.py
requests
timezonefinder
python-dotenv
